refactor(llm-agent): move external-tools-normalizer and tool-call-deltas from llm-agent-server
Move external-tools-normalizer.ts and tool-call-deltas.ts from
packages/llm-agent-server/src/smart-agent/utils/ to top-level of
packages/llm-agent/src/. Replace server files with re-export shims to
maintain backward compatibility and avoid cascading import updates.
